Prevent erratic probe behaviour from causing G29 crashes
When the Z probe is not reliable, it doesn't always trigger when the
bed is hit, or it doesn't always untrigger when the probe is above
the bed. If it simply returns extreme values, a G29 can move the
head outside the print volume (on a delta). With this change
probe_pt returns NaN if the probe failed (or the probe deployment
or stow fails) and G29 recognizes this and stops probing.